feat: Use postgis image and selinux fix
The postgis image is required for some modules on OpenSPP. Like SPP Area
Gis. This is not needed for each project (ie: if we don't use the SPP
Area Gis module), but it is needed to test that
this module can get installed and tested correctly.

Fixed selinux issues on volumes by adding an annotation.
See the [Docker documentation](https://docs.docker.com/engine/storage/bind-mounts/#configure-the-selinux-label) for more information.