Rework the Dockerfile and the Docker bake file (#4)
* feat: rework Dockerfile

The Dockerfile used a supplementary stage that did not add value.
Some of the packages are missing to make OpenSPP work fully (ie:
wkhtmltopdf). The node dependency is only needed for right to left
languages (using rtlcss), and should not be included by default if not needed.
A variable was added for left to right support. The steps were
compressed, to minimize the image size.
Rework the docker-bake file to have better image tagging

* fix: Dockerfile and Dockerfile.slim

Change sha1 to sha256 (security)
Remove bash -x argument
Fix typo (stage 3, not 2)