Improve the shell scripts robustness, security and building process (#3)
* doc: Update README.md

Added a Caution regarding the docker-compose. It is intended for testing
only. Not production. For production, check the OpenSPP documentation.

* feat: improve shell scripts

Fix shellcheck issues and use shfmt for indentation.

* feat: use Docker buildx

Use docker buildx for building the images.
Use --no-cache.
Add the docker bake target.

* feat: Add Docker bake

Add Docker bake for OS matrix building.

* fix: Odoo config file

The queue job part should be at the end of the file, to prevent errors.
The longpolling port was replaced by the gevent port.

* feat: less agressive retries

Changed the number of retries to 24, and the sleep time to 5 seconds.
This makes a total of 120 seconds. The same as previously. Just that on
Cloud, we do not need to test so agressively.

* feat: Use postgis image and selinux fix

The postgis image is required for some modules on OpenSPP. Like SPP Area
Gis. This is not needed for each project (ie: if we don't use the SPP
Area Gis module), but it is needed to test that
this module can get installed and tested correctly.

Fixed selinux issues on volumes by adding an annotation.
See the [Docker documentation](https://docs.docker.com/engine/storage/bind-mounts/#configure-the-selinux-label) for more information.

* Rework the Dockerfile and the Docker bake file (#4)

* feat: rework Dockerfile

The Dockerfile used a supplementary stage that did not add value.
Some of the packages are missing to make OpenSPP work fully (ie:
wkhtmltopdf). The node dependency is only needed for right to left
languages (using rtlcss), and should not be included by default if not needed.
A variable was added for left to right support. The steps were
compressed, to minimize the image size.
Rework the docker-bake file to have better image tagging

* fix: Dockerfile and Dockerfile.slim

Change sha1 to sha256 (security)
Remove bash -x argument
Fix typo (stage 3, not 2)

* [chore] Add makefile comment to the bake command

---------

Co-authored-by: Ken Lewerentz <ken@openspp.org>